Replace parseInt with Math.floor in seed generation
This code selects a random array index from `available` or `byte`. The
formular for this operation is `Math.floor(Math.random()*items.length)`
(see https://stackoverflow.com/a/5915122).

In the current code, Math.floor is replaced with parseInt, which is
meant to convert a string, not a floating point number. Since parseInt
converts a number to a string first, the result will be the same in most
cases. However, parseInt is not the right tool in this place. MDN
excplicitly warns: "parseInt should not be used as a substitute for
Math.floor()"
(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/parseInt)